Comment out default env vars in docker-compose
Since I see the "Port", "BedrockPort", and "BackupCount" variables are already set to their default values, I commented these out since they are redundant. I left a brief note for users on this, like you did for the other commented-out values.

One value I was unable to verify was set to a default (likely it is not) is the TZ variable because I was unable to find where it was actually being used in the entire codebase. Could you explain to me how/where it is used and what would be its default value if left unset?

I see you have it set to America/Denver in each Dockerfile, so if commented out in docker-compose it would take this value, I understand. But what would happen if you didn't set it in the Dockerfiles either?

Signed-off-by: William Trelawny <22324745+willman42@users.noreply.github.com>